Adapt scaling of forward kernels. Add a torch chunkwise fwbw impl.
Squashed commit of the following:

commit 813c6140176cc5a72d8375334eff56464d907586
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Mon Aug 19 10:22:10 2024 +0200

    remove/comment out print statements in torch chunkwise.

commit 71fce12301793751f139223f9faaffd6f5ba24a1
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Mon Aug 19 10:02:24 2024 +0200

    apply qk_scale to q instead of k in inference kernels.

commit afeb79d8d962f6030de773fa912e7ee146d8bb75
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 19:39:08 2024 +0200

    add vecDeltaI.

commit 94b9979b148467b26c6f5bd60ed3496f478951f2
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 19:24:18 2024 +0200

    Add vecDeltaF gradients.

commit e96ecfc458a4355ad457b69fac83155f7dc900b4
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 19:14:44 2024 +0200

    rename.

commit 3005c69da48ee72aa6d17117e2d52e21e757c897
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 19:03:59 2024 +0200

    Finally! backward gradients match. need to use the correct m states.

commit b7246e31d849c182a9eb19821c340e04ef5d6e72
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 18:19:28 2024 +0200

    debug matQ grad. Values match.

commit 8f31b2f40948e9644f594a5287e03c4a1947bc8e
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 18:19:17 2024 +0200

    before q grad debug.

commit d027979d6b534139509c06d849198cfe9522a5a2
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 17:18:01 2024 +0200

    prepare change of scaling.

commit 5ff694f235d45375315a8274fb91c590451b7bba
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 16:44:08 2024 +0200

    adapt qk scale more.

commit 208ab5e9dff7e9ca4be46a845d14a29310bed4a0
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sun Aug 18 08:49:55 2024 +0200

    more debugging, found a bug. 1/sqrt(d) missing.

commit ab9a1e5a8cf916a11e2f01036cb2bb3910972f43
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sat Aug 17 09:53:47 2024 +0200

    more debugging, wip.

commit ced45613a8071d635f1dd9e4becbce9fda17554a
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Sat Aug 17 09:17:41 2024 +0200

    first fix attempt. Values do not match yet for more than one chunks.

commit ff372cdb8b1b56c77a2e4f3ae4fae6c5cb847515
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Fri Aug 16 18:18:23 2024 +0200

    fix forward pass recurrent_fw: computing NC + 1 states (also the last one).

commit 4a9ac44c04682c841e1f859569f864944993315f
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Thu Aug 15 18:18:46 2024 +0200

    dQ dK dV match if there is only one chunk. some bug in recurrent part.

commit f42083508d7e4eceb42368f14b35f4bacbdf4f98
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Thu Aug 15 16:42:30 2024 +0200

    add _mlstm_chunkwise__parallel_bw_dQKV. numerical values do not match yet.

commit 24862b3ff2e5e9c62ef6e740348e755e10d290f3
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Thu Aug 15 14:52:07 2024 +0200

    prepare recurrent to return also the states.

commit a5f85928ad1ad1456885438e16e71f5a00911a1e
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Thu Aug 15 14:29:03 2024 +0200

    adapt interface of recurrent sequence fw. torch version.

commit debed31f9e9d8fb3db6339e78726943530413ed0
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Thu Aug 15 14:04:13 2024 +0200

    add _mlstm_chunkwise__recurrent_bw_dC().

commit 669f4ccabe54a8a667c70f7bbbf903ff7af52454
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Thu Aug 15 12:25:19 2024 +0200

    adapt interface for  _mlstm_chunkwise__recurrent_bw_dC.

commit 597ed710a60cb5540e4b6aab3d37796729511db8
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Thu Aug 15 09:53:16 2024 +0200

    output denominator in forward for backward.

commit 73988f115479a31d6da8e0893bc5f34581719241
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Wed Aug 14 19:52:43 2024 +0200

    add torch fwbw autograd boilerplate code.

commit 6aaf37e45fb0b8dba254ae920e401d7fad193249
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Wed Aug 14 17:12:41 2024 +0200

    adapt naming of mlstm_chunkwise_fw. reorganize in two subfunctions.

commit 2cab3ac8d863b3940b6f154920b12ff6b5653ed1
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Wed Aug 14 10:48:47 2024 +0200

    adapt naming and fix scaling. numerical values match.

commit 49b724514b05c98eb705a12409e15f0c7fa247ce
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Wed Aug 14 08:58:28 2024 +0200

    add an overview over chunkwise kernels.

commit bb41b7e8bf4f1248a856f36aee064f2f799c1fcc
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Wed Aug 14 08:42:36 2024 +0200

    rename.

commit 510ceb15a288eff3bbdac6e9bd74605a390851e3
Author: Maximilian Beck <beck@ml.jku.at>
Date:   Tue Aug 13 15:02:58 2024 +0200

    add todo.

+++ mlstm_kernels/mlstm/parallel/_triton_bw.py
@@ -725,13 +725,11 @@ def mlstm_bw(
 
     ## ? postprocessing
     # compute the vecDeltaFbar values with dfbar = rev_cumsum((q*dq - k*dk).sum(-1))
-    # TODO should we cast to float32 here?
-    vecDeltaFbar_acc = (
-        matQ.to(dtype=torch.float32) * matDeltaQ.to(dtype=torch.float32)
-        - matK.to(dtype=torch.float32) * matDeltaK.to(dtype=torch.float32)
-    ).sum(-1)
+    # should we cast to float32 here?
+    # No! this causes loading and casting all the data again.
+    vecDeltaFbar_acc = (matQ * matDeltaQ - matK * matDeltaK).sum(-1)
     vecDeltaFbar = vecDeltaFbar_acc.flip(-1).cumsum(-1).flip(-1)
-    vecDeltaF = (vecDeltaFbar * torch.sigmoid(-vecF)).to(dtype=vecF.dtype)
+    vecDeltaF = (vecDeltaFbar * torch.sigmoid(-vecF))
     ## ? end postprocessing
 
     return matDeltaQ, matDeltaK, matDeltaV, vecDeltaI, vecDeltaF
